refactored the data package to split between CSV and Excel
