+++ src/main/java/autocadDrawingChecker/data/excel/autoCADData/AutoCADElement.java
@@ -1,4 +1,4 @@
-package autocadDrawingChecker.data.autoCADData;
+package autocadDrawingChecker.data.excel.autoCADData;
 
 import autocadDrawingChecker.data.core.Record;
