+++ src/main/java/autocadDrawingChecker/grading/criteria/autoCADCriteria/AbstractAutoCADElementCriteria.java
@@ -5,8 +5,8 @@
  */
 package autocadDrawingChecker.grading.criteria.autoCADCriteria;
 
-import autocadDrawingChecker.data.autoCADData.AutoCADElement;
-import autocadDrawingChecker.data.autoCADData.AutoCADExport;
+import autocadDrawingChecker.data.excel.autoCADData.AutoCADElement;
+import autocadDrawingChecker.data.excel.autoCADData.AutoCADExport;
 import autocadDrawingChecker.data.core.DataSet;
 import autocadDrawingChecker.data.core.Record;
 import autocadDrawingChecker.grading.criteria.AbstractElementCriteria;
